+++ PrintInterface/PrintTransactionModel .cs
@@ -15,9 +15,9 @@ namespace PrintInterface
         public string TransactionDateCreated;
         public string TransactionDescription = "";
         public string TransactionFinStatus;
-        public double TotalBalance;
-        public double TotalPositiveItemsSum;
-        public double TotalNegativeItemsSum;
+        public decimal TotalBalance;
+        public decimal TotalPositiveItemsSum;
+        public decimal TotalNegativeItemsSum;
 
         public int TransactionType;
         public bool PrintTransactionDescription;
