+++ SharedLibrary/DalModels/ProductModel.cs
@@ -17,6 +17,6 @@ public class ProductModel
     public string TimeUpdated { get; set; }
     public string Descriptions { get; set; }
     public bool IsActive { get; set; } = true;
-    public double CountValue => StandardCalculator.CalculateFromString(CountString);
+    public decimal CountValue => StandardCalculator.CalculateFromString(CountString);
     public bool IsCountStringValid => StandardCalculator.IsCalculatable(CountString);
 }
