+++ SharedLibrary/DalModels/TransactionModel.cs
@@ -16,11 +16,11 @@ public class TransactionModel
     public ObservableCollection<TransactionItemModel> Items { get; set; }
     public string Descriptions { get; set; }
 
-    public double PositiveItemsSum => Items == null ? 0 : Items.Where(i => i.TotalValue > 0).Sum(i => i.TotalValue);
-    public double NegativeItemsSum => Items == null ? 0 : Items.Where(i => i.TotalValue < 0).Sum(i => i.TotalValue);
-    public double Balance => PositiveItemsSum + NegativeItemsSum;
-    public double TotalPositiveItemsSum { get; set; }
-    public double TotalNegativeItemsSum { get; set; }
-    public double TotalBalance => TotalPositiveItemsSum + TotalNegativeItemsSum;
+    public decimal PositiveItemsSum => Items == null ? 0 : Items.Where(i => i.TotalValue > 0).Sum(i => i.TotalValue);
+    public decimal NegativeItemsSum => Items == null ? 0 : Items.Where(i => i.TotalValue < 0).Sum(i => i.TotalValue);
+    public decimal Balance => PositiveItemsSum + NegativeItemsSum;
+    public decimal TotalPositiveItemsSum { get; set; }
+    public decimal TotalNegativeItemsSum { get; set; }
+    public decimal TotalBalance => TotalPositiveItemsSum + TotalNegativeItemsSum;
     public TransactionFinancialStatus TransactionFinancialStatus => Balance == 0 ? TransactionFinancialStatus.Balanced : Balance > 0 ? TransactionFinancialStatus.Positive : TransactionFinancialStatus.Negative;
 }
